style: Ignore testdata files in editorconfig linter
Also make archive related words constants
