+++ pkg/constant/compression.go
@@ -1,0 +1,11 @@
+package constant
+
+// nolint:revive
+const (
+	CompressionMethod_Gzip    = "gzip"
+	CompressionMethod_Bzip2   = "bzip2"
+	CompressionMethod_XZ      = "xz"
+	CompressionMethod_LZMA    = "lzma"
+	CompressionMethod_ZSTD    = "zstd"
+	CompressionMethod_DEFLATE = "deflate"
+)
