+++ pkg/tools/tests/task.go
@@ -1,13 +1,21 @@
 package tests
 
 import (
+	"context"
 	"reflect"
 	"testing"
 
+	"arhat.dev/pkg/testhelper"
 	"arhat.dev/rs"
 	"github.com/stretchr/testify/assert"
 
 	"arhat.dev/dukkha/pkg/dukkha"
+	dukkha_test "arhat.dev/dukkha/pkg/dukkha/test"
+	"arhat.dev/dukkha/pkg/renderer/env"
+	"arhat.dev/dukkha/pkg/renderer/file"
+	"arhat.dev/dukkha/pkg/renderer/shell"
+	"arhat.dev/dukkha/pkg/renderer/template"
+	"arhat.dev/dukkha/pkg/tools"
 )
 
 type ExecSpecGenerationTestCase struct {
@@ -73,4 +81,96 @@ func runTaskTest(taskCtx dukkha.TaskExecContext, test *ExecSpecGenerationTestCas
 	}
 
 	assert.Equal(t, test.Expected, specs)
+}
+
+func TestTask(
+	t *testing.T,
+	dir string,
+	tool dukkha.Tool,
+	newTask func() dukkha.Task,
+	newExpected func() rs.Field,
+	check func(expected, actual rs.Field),
+) {
+	type TestCase struct {
+		rs.BaseField
+
+		// Tool dukkha.Tool `yaml:"tool"`
+		Task dukkha.Task `yaml:"task"`
+	}
+
+	type CheckSpec struct {
+		rs.BaseField
+
+		ExpectErr bool     `yaml:"expect_err"`
+		Actual    rs.Field `yaml:"actual"`
+		Expected  rs.Field `yaml:"expected"`
+	}
+
+	testhelper.TestFixtures(t, dir,
+		func() interface{} {
+			return rs.Init(&TestCase{}, &rs.Options{
+				InterfaceTypeHandler: rs.InterfaceTypeHandleFunc(
+					func(typ reflect.Type, yamlKey string) (interface{}, error) {
+						return rs.Init(newTask(), nil), nil
+					},
+				),
+			})
+		},
+		func() interface{} {
+			return rs.Init(&CheckSpec{}, &rs.Options{
+				InterfaceTypeHandler: rs.InterfaceTypeHandleFunc(
+					func(typ reflect.Type, yamlKey string) (interface{}, error) {
+						return rs.Init(newExpected(), nil), nil
+					},
+				),
+			})
+		},
+		func(t *testing.T, in, exp interface{}) {
+			defer t.Cleanup(func() {
+
+			})
+			spec := in.(*TestCase)
+			e := exp.(*CheckSpec)
+
+			ctx := dukkha_test.NewTestContext(context.TODO())
+			ctx.SetCacheDir(t.TempDir())
+			ctx.AddRenderer("file", file.NewDefault("file"))
+			ctx.AddRenderer("env", env.NewDefault("env"))
+			ctx.AddRenderer("template", template.NewDefault("template"))
+			ctx.AddRenderer("shell", shell.NewDefault("shell"))
+
+			if !assert.NoError(t, spec.ResolveFields(ctx, -1)) {
+				return
+			}
+
+			rs.Init(tool, nil)
+
+			assert.NoError(t, tool.Init("", ctx.CacheDir()))
+			ctx.AddTool(tool.Key(), tool)
+
+			assert.NoError(t, tool.AddTasks([]dukkha.Task{spec.Task}))
+
+			err := tools.RunTask(&tools.TaskExecRequest{
+				Context:     ctx,
+				Tool:        tool,
+				Task:        spec.Task,
+				IgnoreError: false,
+			})
+
+			if !assert.NoError(t, e.ResolveFields(ctx, -1)) {
+				return
+			}
+
+			if e.ExpectErr {
+				assert.Error(t, err)
+				return
+			}
+
+			if !assert.NoError(t, err) {
+				return
+			}
+
+			check(e.Expected, e.Actual)
+		},
+	)
 }
